Research_AI_law/F5_F8_MVP/FASE6/pyproject.toml
[project]
name = "fase6-modeling"
version = "0.4.0"
description = "Fase 6 — Modelado MVP: asociación regulación IA vs ecosistema"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.0",
  "numpy>=1.24",
  "scipy>=1.11",
  "scikit-learn>=1.3",
  "statsmodels>=0.14",
  "pyyaml>=6.0",
  "pytest>=7.4",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
